User is able to input their first and last name.
The user is able to input their first and last name in the input fields. After pressing the submit button, the panel will check if the fields are not empty. If there not empty then the name will be saved in the UI manager. Otherwise a error message will pop up.