Location Panel Finished + Google API
I successfully integrated the Google Map Static API. When the user is on the location panel, if the device allows them to give up their location, the panel will find the x and y coordinates for the map. The program will now download the google map from a unity web request based on the location and display it to the user. The user also has the option to have location notes.